requirements.txt
Flask==3.1.0
supabase==2.15.0
gunicorn==23.0.0
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-auth==2.37.0
gspread==6.1.4
beautifulsoup4==4.12.3
python-dotenv==1.0.1
anthropic==0.42.0
reportlab==4.1.0
flask-limiter==4.1.1
flask-compress==1.17
cryptography==44.0.1
PyJWT==2.10.1
requests==2.32.3
cloudscraper==1.2.71
